Put uclamp vmlinux_deps into their own file
If uclamp is not compiled, then pahole will hit an error because it
won't be able to find the struct definition of uclamp_bucket for
instance as it was compiled out. By separating the deps of uclamp, we
ensure the essential definitions are always created even if
vmlinux_deps_uclamp.h is empty. The sched_tp code is properly protected
by #ifdef CONFIG_UCLAMP_TASK; so it won't try to use it if uclamp is not
enabled.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>